:sparkles: New Button Style
Played around with link styles for this update. Currently they all link to Google, but I'm going to get started on building some of the template pages that I will base pretty much all of the other ones on!

Also: Finally got around to properly adding in Raleway into the files, so now the light and bold varients are on the site.